refactor(ws_server): split into submodules (auth, session, http)
Split the 804-line ws_server.rs into focused submodules:
- mod.rs: public API, server orchestration, binding helpers
- auth.rs: token/origin validation, rate limiting, token generation
- session.rs: WebSocket session loop, command dispatch
- http.rs: raw HTTP helpers (/health, error responses)

Removed low-signal unit tests (rate_entry field checks,
health_json_has_required_fields) that tested struct fields
rather than behavior.
