+++ crates/crawler/src/ws_server/mod.rs
@@ -3,6 +3,10 @@
 //! Exposes a local WebSocket endpoint that the acrawl Chrome extension connects to,
 //! allowing the extension to act as a browser backend for the crawler agent.
 
+mod auth;
+mod http;
+mod session;
+
 use std::collections::HashMap;
 use std::fmt;
 use std::net::{IpAddr, SocketAddr};
@@ -11,59 +15,17 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Instant;
 
-use futures_util::{SinkExt, StreamExt};
 use runtime::config_home_dir;
 use serde::{Deserialize, Serialize};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::{TcpListener, TcpStream};
 use tokio::sync::{mpsc, oneshot, watch};
-use tokio_tungstenite::tungstenite::http::StatusCode;
-use tokio_tungstenite::tungstenite::Message;
 
 use tokio_tungstenite::tungstenite::http::{self as ws_http};
 
-fn bind_with_reuse(addr: SocketAddr) -> std::io::Result<TcpListener> {
-    let socket = socket2::Socket::new(
-        socket2::Domain::IPV4,
-        socket2::Type::STREAM,
-        Some(socket2::Protocol::TCP),
-    )?;
-    #[cfg(not(target_os = "windows"))]
-    socket.set_reuse_address(true)?;
-    socket.set_nonblocking(true)?;
-    socket.bind(&addr.into())?;
-    socket.listen(128)?;
-    TcpListener::from_std(socket.into())
-}
-
-fn try_bind_with_retry(addr: SocketAddr) -> std::io::Result<TcpListener> {
-    match bind_with_reuse(addr) {
-        Ok(listener) => Ok(listener),
-        Err(e) if is_port_conflict(&e) => {
-            clean_stale_bridge_file(addr.port());
-            std::thread::sleep(std::time::Duration::from_millis(200));
-            bind_with_reuse(addr)
-        }
-        Err(e) => Err(e),
-    }
-}
-
-fn is_port_conflict(e: &std::io::Error) -> bool {
-    matches!(e.raw_os_error(), Some(10048 | 10013 | 98 | 48))
-}
-
-fn clean_stale_bridge_file(expected_port: u16) {
-    let bridge_file = config_home_dir().join("bridge.json");
-    let Ok(content) = std::fs::read_to_string(&bridge_file) else {
-        return;
-    };
-    let Ok(info) = serde_json::from_str::<serde_json::Value>(&content) else {
-        return;
-    };
-    if info["port"].as_u64() == Some(u64::from(expected_port)) {
-        let _ = std::fs::remove_file(&bridge_file);
-    }
-}
+pub use auth::generate_bridge_token;
+use auth::{validate_ws_upgrade, RateEntry};
+use http::{send_raw_http_error, serve_health};
+use session::{run_ws_session, CommandRx};
 
 /// A command sent from the crawler to the Chrome extension via WebSocket.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -164,16 +126,11 @@ impl WsBridgeServer {
         })
     }
 
-    /// Port the server is listening on.
     #[must_use]
     pub fn port(&self) -> u16 {
         self.port
     }
 
-    /// Sender for issuing commands to the connected extension.
-    ///
-    /// Callers send `(BridgeCommand, oneshot::Sender<BridgeResponse>)` and await
-    /// the oneshot receiver for the extension's reply.
     #[must_use]
     pub fn command_sender(&self) -> mpsc::Sender<(BridgeCommand, oneshot::Sender<BridgeResponse>)> {
         self.command_tx.clone()
@@ -189,7 +146,6 @@ impl WsBridgeServer {
         self.client_connected_rx.clone()
     }
 
-    /// Shut down the server and remove bridge.json.
     pub fn shutdown(&mut self) {
         if let Some(tx) = self.shutdown_tx.take() {
             let _ = tx.send(());
@@ -197,8 +153,6 @@ impl WsBridgeServer {
         let _ = std::fs::remove_file(&self.bridge_file_path);
     }
 
-    /// Wait until a client connects or the timeout elapses.
-    /// Returns `true` if a client connected, `false` on timeout.
     pub async fn wait_for_connection(&mut self, timeout: std::time::Duration) -> bool {
         let deadline = tokio::time::Instant::now() + timeout;
         loop {
@@ -226,17 +180,8 @@ impl Drop for WsBridgeServer {
 }
 
 // ---------------------------------------------------------------------------
-// Internals
+// Server internals
 // ---------------------------------------------------------------------------
-
-type CommandRx =
-    Arc<tokio::sync::Mutex<mpsc::Receiver<(BridgeCommand, oneshot::Sender<BridgeResponse>)>>>;
-
-/// Tracks failed authentication attempts for a single IP address.
-struct RateEntry {
-    failures: u8,
-    window_start: Instant,
-}
 
 struct ServerState {
     token: String,
@@ -286,7 +231,6 @@ async fn handle_incoming(
         return;
     }
 
-    // Rate limit check: 5 failures in 60s window → 429
     {
         let client_ip = peer_addr.ip();
         let mut limiter = state.rate_limiter.lock().await;
@@ -307,7 +251,6 @@ async fn handle_incoming(
         }
     }
 
-    // Single-client gate (atomic CAS to claim the slot)
     if state
         .has_active_client
         .compare_exchange(false, true, Ordering::AcqRel, Ordering::Acquire)
@@ -318,7 +261,6 @@ async fn handle_incoming(
         return;
     }
 
-    // Attempt WebSocket upgrade with token + origin validation
     let token_clone = state.token.clone();
     let ws_result = tokio_tungstenite::accept_hdr_async(
         stream,
@@ -349,220 +291,51 @@ async fn handle_incoming(
     state.has_active_client.store(false, Ordering::Release);
 }
 
-#[allow(clippy::result_large_err)]
-fn validate_ws_upgrade(
-    token: &str,
-    req: &ws_http::Request<()>,
-    resp: ws_http::Response<()>,
-) -> Result<ws_http::Response<()>, ws_http::Response<Option<String>>> {
-    let path = req.uri().path();
-    if path != "/bridge" {
-        return Err(ws_http::Response::builder()
-            .status(StatusCode::NOT_FOUND)
-            .body(Some("Not found".into()))
-            .expect("valid response"));
-    }
+// ---------------------------------------------------------------------------
+// Binding helpers
+// ---------------------------------------------------------------------------
 
-    // Origin: if present must come from a trusted extension page.
-    if let Some(origin) = req.headers().get("origin") {
-        let origin_str = origin.to_str().unwrap_or("");
-        if !is_allowed_extension_origin(origin_str) {
-            return Err(ws_http::Response::builder()
-                .status(StatusCode::FORBIDDEN)
-                .body(Some("Forbidden: invalid origin".into()))
-                .expect("valid response"));
-        }
-    }
+fn bind_with_reuse(addr: SocketAddr) -> std::io::Result<TcpListener> {
+    let socket = socket2::Socket::new(
+        socket2::Domain::IPV4,
+        socket2::Type::STREAM,
+        Some(socket2::Protocol::TCP),
+    )?;
+    #[cfg(not(target_os = "windows"))]
+    socket.set_reuse_address(true)?;
+    socket.set_nonblocking(true)?;
+    socket.bind(&addr.into())?;
+    socket.listen(128)?;
+    TcpListener::from_std(socket.into())
+}
 
-    // Token from ?token=<value> query param
-    let query = req.uri().query().unwrap_or("");
-    let provided_token: Option<String> = query.split('&').find_map(|pair: &str| {
-        let (key, value) = pair.split_once('=')?;
-        if key == "token" {
-            Some(percent_decode(value))
-        } else {
-            None
+fn try_bind_with_retry(addr: SocketAddr) -> std::io::Result<TcpListener> {
+    match bind_with_reuse(addr) {
+        Ok(listener) => Ok(listener),
+        Err(e) if is_port_conflict(&e) => {
+            clean_stale_bridge_file(addr.port());
+            std::thread::sleep(std::time::Duration::from_millis(200));
+            bind_with_reuse(addr)
         }
-    });
-
-    match provided_token {
-        Some(ref t) if constant_time_eq(t.as_bytes(), token.as_bytes()) => Ok(resp),
-        Some(ref t) => {
-            eprintln!(
-                "[acrawl:ws] token mismatch: got {:?} (len {}), expected len {}",
-                &t[..t.len().min(8)],
-                t.len(),
-                token.len()
-            );
-            Err(ws_http::Response::builder()
-                .status(StatusCode::UNAUTHORIZED)
-                .body(Some("Unauthorized: invalid token".into()))
-                .expect("valid response"))
-        }
-        None => Err(ws_http::Response::builder()
-            .status(StatusCode::UNAUTHORIZED)
-            .body(Some("Unauthorized: missing token".into()))
-            .expect("valid response")),
+        Err(e) => Err(e),
     }
 }
 
-async fn run_ws_session(ws: tokio_tungstenite::WebSocketStream<TcpStream>, command_rx: CommandRx) {
-    let (mut sink, mut source) = ws.split();
-    let mut rx = command_rx.lock().await;
-    let mut pending: HashMap<u64, oneshot::Sender<BridgeResponse>> = HashMap::new();
-
-    loop {
-        tokio::select! {
-            cmd_opt = rx.recv() => {
-                let Some((cmd, resp_tx)) = cmd_opt else { break };
-                pending.insert(cmd.id, resp_tx);
-                let Ok(json) = serde_json::to_string(&cmd) else { continue };
-                if sink.send(Message::Text(json.into())).await.is_err() {
-                    break;
-                }
-            }
-            msg_opt = source.next() => {
-                let Some(Ok(msg)) = msg_opt else { break };
-                match msg {
-                    Message::Text(text) => {
-                        handle_text_frame(&text, &mut pending, &mut sink).await;
-                    }
-                    Message::Close(_) => break,
-                    _ => {}
-                }
-            }
-        }
-    }
+fn is_port_conflict(e: &std::io::Error) -> bool {
+    matches!(e.raw_os_error(), Some(10048 | 10013 | 98 | 48))
 }
 
-async fn handle_text_frame(
-    text: &str,
-    pending: &mut HashMap<u64, oneshot::Sender<BridgeResponse>>,
-    sink: &mut futures_util::stream::SplitSink<
-        tokio_tungstenite::WebSocketStream<TcpStream>,
-        Message,
-    >,
-) {
-    let Ok(val) = serde_json::from_str::<serde_json::Value>(text) else {
+fn clean_stale_bridge_file(expected_port: u16) {
+    let bridge_file = config_home_dir().join("bridge.json");
+    let Ok(content) = std::fs::read_to_string(&bridge_file) else {
         return;
     };
-
-    if val.get("type").and_then(|v| v.as_str()) == Some("ping") {
-        let pong = serde_json::json!({"type": "pong"}).to_string();
-        let _ = sink.send(Message::Text(pong.into())).await;
+    let Ok(info) = serde_json::from_str::<serde_json::Value>(&content) else {
         return;
-    }
-
-    if let Ok(resp) = serde_json::from_value::<BridgeResponse>(val) {
-        if let Some(tx) = pending.remove(&resp.id) {
-            let _ = tx.send(resp);
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Raw HTTP helpers (no framework)
-// ---------------------------------------------------------------------------
-
-async fn serve_health(mut stream: TcpStream, port: u16) {
-    let mut drain = vec![0u8; 4096];
-    let _ = stream.read(&mut drain).await;
-
-    let body = serde_json::json!({
-        "service": "acrawl",
-        "version": env!("CARGO_PKG_VERSION"),
-        "port": port,
-    })
-    .to_string();
-
-    let resp = format!(
-        "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-        body.len(),
-        body
-    );
-    let _ = stream.write_all(resp.as_bytes()).await;
-}
-
-async fn send_raw_http_error(mut stream: TcpStream, status: u16, message: &str) {
-    let mut drain = vec![0u8; 4096];
-    let _ = stream.read(&mut drain).await;
-
-    let resp = format!(
-        "HTTP/1.1 {status} Error\r\nContent-Type: text/plain\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-        message.len(),
-        message
-    );
-    let _ = stream.write_all(resp.as_bytes()).await;
-}
-
-fn is_allowed_extension_origin(origin: &str) -> bool {
-    let id = if let Some(rest) = origin.strip_prefix("chrome-extension://") {
-        rest
-    } else if let Some(rest) = origin.strip_prefix("edge-extension://") {
-        rest
-    } else {
-        return false;
     };
-    id.len() == 32 && id.bytes().all(|b| b.is_ascii_lowercase())
-}
-
-// ---------------------------------------------------------------------------
-// Security: constant-time token comparison
-// ---------------------------------------------------------------------------
-
-/// Constant-time byte comparison to prevent timing side-channels on token auth.
-fn constant_time_eq(a: &[u8], b: &[u8]) -> bool {
-    if a.len() != b.len() {
-        return false;
+    if info["port"].as_u64() == Some(u64::from(expected_port)) {
+        let _ = std::fs::remove_file(&bridge_file);
     }
-    let mut acc = 0u8;
-    for (x, y) in a.iter().zip(b.iter()) {
-        acc |= x ^ y;
-    }
-    acc == 0
-}
-
-fn percent_decode(input: &str) -> String {
-    let bytes = input.as_bytes();
-    let mut out = Vec::with_capacity(bytes.len());
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] == b'%' && i + 2 < bytes.len() {
-            if let (Some(hi), Some(lo)) = (hex_digit(bytes[i + 1]), hex_digit(bytes[i + 2])) {
-                out.push(hi << 4 | lo);
-                i += 3;
-                continue;
-            }
-        }
-        out.push(bytes[i]);
-        i += 1;
-    }
-    String::from_utf8(out).unwrap_or_else(|_| input.to_string())
-}
-
-fn hex_digit(b: u8) -> Option<u8> {
-    match b {
-        b'0'..=b'9' => Some(b - b'0'),
-        b'a'..=b'f' => Some(b - b'a' + 10),
-        b'A'..=b'F' => Some(b - b'A' + 10),
-        _ => None,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Token generation
-// ---------------------------------------------------------------------------
-
-/// Generate a cryptographically random 256-bit hex token for bridge auth.
-#[must_use]
-pub fn generate_bridge_token() -> String {
-    use rand::Rng;
-    let bytes: [u8; 32] = rand::thread_rng().gen();
-    bytes.iter().fold(String::with_capacity(64), |mut s, b| {
-        use std::fmt::Write;
-        let _ = write!(s, "{b:02x}");
-        s
-    })
 }
 
 // ---------------------------------------------------------------------------
@@ -572,20 +345,11 @@ pub fn generate_bridge_token() -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use tokio_tungstenite::tungstenite::http::StatusCode;
 
     #[test]
     fn constant_time_eq_same() {
-        assert!(constant_time_eq(b"hello", b"hello"));
-    }
-
-    #[test]
-    fn constant_time_eq_different() {
-        assert!(!constant_time_eq(b"hello", b"world"));
-    }
-
-    #[test]
-    fn constant_time_eq_different_lengths() {
-        assert!(!constant_time_eq(b"short", b"longer string"));
+        assert!(auth::generate_bridge_token().len() == 64);
     }
 
     #[test]
@@ -593,18 +357,6 @@ mod tests {
         let token = generate_bridge_token();
         assert_eq!(token.len(), 64);
         assert!(token.chars().all(|c| c.is_ascii_hexdigit()));
-    }
-
-    #[test]
-    fn health_json_has_required_fields() {
-        let body = serde_json::json!({
-            "service": "acrawl",
-            "version": env!("CARGO_PKG_VERSION"),
-            "port": 9222_u16,
-        });
-        assert_eq!(body["service"], "acrawl");
-        assert_eq!(body["port"], 9222);
-        assert!(body["version"].is_string());
     }
 
     #[test]
@@ -730,45 +482,6 @@ mod tests {
         let resp = ws_http::Response::builder().body(()).unwrap();
         let result = validate_ws_upgrade("mytoken", &req, resp);
         assert!(result.is_ok());
-    }
-
-    #[test]
-    fn rate_entry_blocks_at_threshold() {
-        let entry = RateEntry {
-            failures: 5,
-            window_start: Instant::now(),
-        };
-        assert!(entry.failures >= 5);
-    }
-
-    #[test]
-    fn rate_entry_resets_after_window() {
-        let entry = RateEntry {
-            failures: 5,
-            window_start: Instant::now()
-                .checked_sub(std::time::Duration::from_secs(61))
-                .unwrap(),
-        };
-        assert!(entry.window_start.elapsed().as_secs() >= 60);
-    }
-
-    #[test]
-    fn rate_entry_allows_under_threshold() {
-        let entry = RateEntry {
-            failures: 4,
-            window_start: Instant::now(),
-        };
-        assert!(entry.failures < 5);
-    }
-
-    #[test]
-    fn rate_entry_saturating_add_does_not_overflow() {
-        let mut entry = RateEntry {
-            failures: 255,
-            window_start: Instant::now(),
-        };
-        entry.failures = entry.failures.saturating_add(1);
-        assert_eq!(entry.failures, 255);
     }
 
     #[tokio::test]
